Re-committing previous commits as one
There was something odd going on with the api folder, so redoing the
changes:

* libdeng2: Added function_cast for casting pointers to function pointers

* Refactor: Made linedef_s an opaque type, moved members to LineDef class

This warranted switching almost all engine sources to .cpp.
Also sorted out headers for successfully compiling game plugins.

+++ doomsday/engine/include/gl/gl_main.h
@@ -58,7 +58,7 @@ extern "C" {
     GLint p; \
     glGetIntegerv(GL_TEXTURE_BINDING_2D, &p); \
     Sys_GLCheckError(); \
-    assert(p == tex); \
+    assert(p == (GLint)tex); \
 }
 #else
 #  define LIBDENG_ASSERT_GL_TEXTURE_ISBOUND(tex)
